requirements.txt
pydantic==1.8.2
requests==2.23.0
python-dotenv==0.19.2
joblib==1.1.1
pypandoc==1.5
pyspark==2.4.7
pyyaml==6.0
python-dateutil==2.8.2
lightgbm==3.3.5
pandas==1.1.5
xgboost==1.5.2
colorlog==6.7.0
redis==4.6.0
attrs==23.1.0
pytest==7.4.0

